Extraccion/requirements.txt
gdown
selenium
pandas
openpyxl
sqlalchemy
psycopg2-binary